payment script

+++ common.sh
@@ -43,6 +43,8 @@ else
 ---------------------------------------------------------------------------------------
 pre_reqiuisites1() {
   print_heading1 "add application user"
+   id roboshop &>>$log_file1
+if [$? -ne 0 ]; then
   useradd roboshop &>>$log_file1
   status_check1 $?
   rm -rf /app &>>$log_file1
